Internal command should capture all it's stdout,err (#510)
Using a job hook, we can intercet and buffer all job output
for usage in internal commands where catpuring the full output is important
that's beside the job internal buffer which is limited to only 100 lines.

This means that calling cl.system('command') will return only 100 lines for
each of stdout, and stderr. while doing the same in core0 side (for internal commands)
will capture the full output for further processing.