added isStandby in builtins for a disk
When 0-OS ahs issued a command to put a disk in standby mode after a
specified inactive time, we need to make the monitoring tools aware of
that fact. if the monitoring tools are oblivious of the fact that the
disk is spun down, every read and/or write to that disk will reactivate
that disk and it will be spun up again.